+++ src/game/player/elements/image/AspectScaleImage.tsx
@@ -6,24 +6,25 @@ import {useGame} from "@core/common/player";
 export default function AspectScaleImage(
     {
         ref,
-        onWidthChange,
+        onSizeChanged,
         autoFit = false,
     }: Readonly<{
         ref?: React.RefObject<HTMLImageElement | null>;
-        onWidthChange?: (width: number) => void;
+        onSizeChanged?: (width: number, height: number) => void;
         autoFit?: boolean;
     }>
 ) {
     const imgRef = useRef<HTMLImageElement>(null);
     const {ratio} = useRatio();
     const [width, setWidth] = React.useState<number>(0);
+    const [height, setHeight] = React.useState<number>(0);
     const {game} = useGame();
 
     useEffect(() => {
         updateWidth();
 
         return ratio.onUpdate(updateWidth);
-    }, []);
+    }, [ref]);
 
     useEffect(() => {
         const observer = new MutationObserver((mutations) => {
@@ -47,12 +48,16 @@ export default function AspectScaleImage(
 
     function updateWidth() {
         const currentRef = ref || imgRef;
-        if (currentRef.current) {
+        if (currentRef.current && currentRef.current.naturalWidth) {
             const autoFitFactor = autoFit ? (game.config.player.width / currentRef.current.naturalWidth) : 1;
             const newWidth = currentRef.current.naturalWidth * ratio.state.scale * autoFitFactor;
+            const newHeight = currentRef.current.naturalHeight * ratio.state.scale * autoFitFactor;
+
             setWidth(newWidth);
-            if (onWidthChange) {
-                onWidthChange(newWidth);
+            setHeight(newHeight);
+
+            if (onSizeChanged) {
+                onSizeChanged(newWidth, newHeight);
             }
         }
     }
@@ -66,6 +71,7 @@ export default function AspectScaleImage(
             ref={ref || imgRef}
             onLoad={handleOnLoad}
             width={width}
+            height={height}
             alt={"image"}
         />
     );